requirements.txt
fastapi[standard]==0.116.1
sqlmodel==0.0.24
python-dotenv==1.1.1
psycopg2==2.9.10
pytest==8.4.1